fix(scripts): read confirmation from tty
Reading from stdin doesn't work when install command is behind pipe. In
that case stdin is not tty but output of previous command, which in case
of the installation is curl with the script itself, and nothing more.

The script doesn't have anything to read in this case and immediately
exits without doing anything.

Can be tested by using old and new version with `cat install.sh | bash`